top-level suites: lazily initialize compartments
Fixes an initialization order bug leading to tests not being
executed if a file contains entry point code, and this file also
has a top-level suite property with a compartment assignment.

Before the entry point function is invoked, this files 'Kt' class will
be instantiated with all the file's top-level properties. At this
point, there is no valid `TestSession` object to initialize
compartments.

If the compartment was not initialized lazily, the file initialization
would fail.